fixed bugs

+++ dataset.py
@@ -141,7 +141,7 @@ def collate_fn2(data): #, history_word_num=60
         indication_prompts[i,:len(tokens)] = tokens
         
     indication_prompts = torch.tensor(indication_prompts).type(torch.LongTensor)
-    probs = torch.tensor(probs) 
+    probs = torch.tensor(probs).type(torch.LongTensor)
     targets = torch.tensor(targets).type(torch.LongTensor)
     
     #print(type(prompt), type(label), type(target), type(prob))
